+++ cloud/__init__.py
@@ -1,2 +1,3 @@
 from . import authenticate
 from . import drive
+from . import clean